+++ public/js/tmsm-aquatonic-course-booking-public.js
@@ -806,8 +806,7 @@ var TmsmAquatonicCourseApp = TmsmAquatonicCourseApp || {};
 
   // Disable PDF link after click
   $('#tmsm-aquatonic-course-booking-download-pdf').on('click', function (e) {
-    console.log('click #tmsm-aquatonic-course-booking-download-pdf');
-    $(this).addClass('disabled');
+    //$(this).addClass('disabled');
     $(this).text(TmsmAquatonicCourseApp.i18n.downloading);
   });